+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/service/operation/DriverInfoCrudOperation.java
@@ -5,17 +5,19 @@ import com.dwarfeng.judge.stack.cache.DriverInfoCache;
 import com.dwarfeng.judge.stack.cache.EnabledDriverInfoCache;
 import com.dwarfeng.judge.stack.dao.DriverInfoDao;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes;
-import com.dwarfeng.subgrade.sdk.service.custom.operation.CrudOperation;
+import com.dwarfeng.subgrade.sdk.service.custom.operation.BatchCrudOperation;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 @Component
-public class DriverInfoCrudOperation implements CrudOperation<LongIdKey, DriverInfo> {
+public class DriverInfoCrudOperation implements BatchCrudOperation<LongIdKey, DriverInfo> {
 
     @Autowired
     private DriverInfoDao driverInfoDao;
@@ -26,7 +28,7 @@ public class DriverInfoCrudOperation implements CrudOperation<LongIdKey, DriverI
     private EnabledDriverInfoCache enabledDriverInfoCache;
 
     @Value("${cache.timeout.entity.driver_info}")
-    private long driveTimeout;
+    private long driverInfoTimeout;
 
     @Override
     public boolean exists(LongIdKey key) throws Exception {
@@ -42,7 +44,7 @@ public class DriverInfoCrudOperation implements CrudOperation<LongIdKey, DriverI
                 throw new ServiceException(ServiceExceptionCodes.ENTITY_NOT_EXIST);
             }
             DriverInfo driverInfo = driverInfoDao.get(key);
-            driverInfoCache.push(driverInfo, driveTimeout);
+            driverInfoCache.push(driverInfo, driverInfoTimeout);
             return driverInfo;
         }
     }
@@ -53,7 +55,7 @@ public class DriverInfoCrudOperation implements CrudOperation<LongIdKey, DriverI
             enabledDriverInfoCache.delete(driverInfo.getSectionKey());
         }
 
-        driverInfoCache.push(driverInfo, driveTimeout);
+        driverInfoCache.push(driverInfo, driverInfoTimeout);
         return driverInfoDao.insert(driverInfo);
     }
 
@@ -68,7 +70,7 @@ public class DriverInfoCrudOperation implements CrudOperation<LongIdKey, DriverI
             enabledDriverInfoCache.delete(driverInfo.getSectionKey());
         }
 
-        driverInfoCache.push(driverInfo, driveTimeout);
+        driverInfoCache.push(driverInfo, driverInfoTimeout);
         driverInfoDao.update(driverInfo);
     }
 
@@ -81,5 +83,52 @@ public class DriverInfoCrudOperation implements CrudOperation<LongIdKey, DriverI
 
         driverInfoDao.delete(key);
         driverInfoCache.delete(key);
+    }
+
+    @Override
+    public boolean allExists(List<LongIdKey> keys) throws Exception {
+        return driverInfoCache.allExists(keys) || driverInfoDao.allExists(keys);
+    }
+
+    @Override
+    public boolean nonExists(List<LongIdKey> keys) throws Exception {
+        return driverInfoCache.nonExists(keys) && driverInfoCache.nonExists(keys);
+    }
+
+    @Override
+    public List<DriverInfo> batchGet(List<LongIdKey> keys) throws Exception {
+        if (driverInfoCache.allExists(keys)) {
+            return driverInfoCache.batchGet(keys);
+        } else {
+            if (!driverInfoDao.allExists(keys)) {
+                throw new ServiceException(ServiceExceptionCodes.ENTITY_NOT_EXIST);
+            }
+            List<DriverInfo> driverInfos = driverInfoDao.batchGet(keys);
+            driverInfoCache.batchPush(driverInfos, driverInfoTimeout);
+            return driverInfos;
+        }
+    }
+
+    @Override
+    public List<LongIdKey> batchInsert(List<DriverInfo> driverInfos) throws Exception {
+        List<LongIdKey> keys = new ArrayList<>();
+        for (DriverInfo driverInfo : driverInfos) {
+            keys.add(insert(driverInfo));
+        }
+        return keys;
+    }
+
+    @Override
+    public void batchUpdate(List<DriverInfo> driverInfos) throws Exception {
+        for (DriverInfo driverInfo : driverInfos) {
+            update(driverInfo);
+        }
+    }
+
+    @Override
+    public void batchDelete(List<LongIdKey> keys) throws Exception {
+        for (LongIdKey key : keys) {
+            delete(key);
+        }
     }
 }