+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/service/operation/SectionCrudOperation.java
@@ -10,7 +10,7 @@ import com.dwarfeng.judge.stack.dao.SectionDao;
 import com.dwarfeng.judge.stack.service.DriverInfoMaintainService;
 import com.dwarfeng.judge.stack.service.JudgerInfoMaintainService;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes;
-import com.dwarfeng.subgrade.sdk.service.custom.operation.CrudOperation;
+import com.dwarfeng.subgrade.sdk.service.custom.operation.BatchCrudOperation;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +21,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Component
-public class SectionCrudOperation implements CrudOperation<LongIdKey, Section> {
+public class SectionCrudOperation implements BatchCrudOperation<LongIdKey, Section> {
 
     @Autowired
     private SectionDao sectionDao;
@@ -101,5 +101,48 @@ public class SectionCrudOperation implements CrudOperation<LongIdKey, Section> {
 
         sectionDao.delete(key);
         sectionCache.delete(key);
+    }
+
+    @Override
+    public boolean allExists(List<LongIdKey> keys) throws Exception {
+        return sectionCache.allExists(keys) || sectionDao.allExists(keys);
+    }
+
+    @Override
+    public boolean nonExists(List<LongIdKey> keys) throws Exception {
+        return sectionCache.nonExists(keys) && sectionCache.nonExists(keys);
+    }
+
+    @Override
+    public List<Section> batchGet(List<LongIdKey> keys) throws Exception {
+        if (sectionCache.allExists(keys)) {
+            return sectionCache.batchGet(keys);
+        } else {
+            if (!sectionDao.allExists(keys)) {
+                throw new ServiceException(ServiceExceptionCodes.ENTITY_NOT_EXIST);
+            }
+            List<Section> sections = sectionDao.batchGet(keys);
+            sectionCache.batchPush(sections, sectionTimeout);
+            return sections;
+        }
+    }
+
+    @Override
+    public List<LongIdKey> batchInsert(List<Section> sections) throws Exception {
+        sectionCache.batchPush(sections, sectionTimeout);
+        return sectionDao.batchInsert(sections);
+    }
+
+    @Override
+    public void batchUpdate(List<Section> sections) throws Exception {
+        sectionCache.batchPush(sections, sectionTimeout);
+        sectionDao.batchUpdate(sections);
+    }
+
+    @Override
+    public void batchDelete(List<LongIdKey> keys) throws Exception {
+        for (LongIdKey key : keys) {
+            delete(key);
+        }
     }
 }